api/requirements.txt
flask
flask-cors
youtube-transcript-api
google-genai
youtube-search
